Add a symlink to gnome-file-roller.tap to fix file-roller detection.
Some distributions (Fedora at least) seem to rename file-roller.desktop
to gnome-file-roller.desktop. The archiver detection in TPA uses the
desktop entry prefix (e.g. file-roller for file-roller.desktop) and
matches this against the .tap file prefixes. So if there only is a
gnome-file-roller.desktop we also need to have a gnome-file-roller.tap.

We therefor install a symlink to work around this.