preserve callbacks in tap-provider
The callbacks are already set to a function which itself calls g_spawn_close_pid.
There is apparently no need to reset the callback to something else.
Either the callback is called, then g_spawn_close_pid will run as intented.
Or nothing will run the callback, in which case nothing will happen anyway.

Maybe all of this was required in 2006.

Signed-off-by: Olaf Hering <olaf@aepfle.de>