Use recently used location for Extract-to dialog
With this change the path in the Extract-to dialog is taken
from the last used location. It is useful if the user wants
to extract multiple archives in the same folder placed deep in
the filesystem tree. The idea is to add similar to Windows
archivers behaviour to improve the user experience when working
with archives.
Works as expected with engrampa and ark.
File-roller opens the last used destination correctly as well,
but since it writes the archive path to GtkRecent, the second
call of Extract-to command will open the location of the archive.
Changes:
- update arguments for engrampa
- add missing destination argument for ark
- fix arguments format for file-roller
- use GtkRecentManager to determine the recently used location
- fallback to default behavior in case of error