Use recently used location for Extract-to dialog
With this change the path in the Extract-to dialog is taken
from the last used location. It is useful if the user wants
to extract multiple archives to the same folder placed deep in
the filesystem tree.
Works as expected with engrampa and ark.
File-roller has a bit different behavior since it writes the
last used archive path to GtkRecent instead of the extracted path,
so the path is not reloaded by the next Extract-to dialog call.