Add destination selection dialog for Extract-to operation
This commit introduces destination location selection dialog.
Dialog is opened on the recently used location by default.

If the User selects a new destination it is passed to the
archive-manager which may or may not open its own dialog
window for confirmation depends on the implementation.

This feature is useful when the user wants to extract multiple archives
to the same folder placed deep in the filesystem tree.

Signed-off-by: Sergey Kovalenko <seryoga.engineering@gmail.com>